--- common/name_generator/data_provider/nature_theme_data.go
@@ -1,257 +1,0 @@
-package data_provider
-
-// GetNatureThemeNounsData it is a set so that we do not have duplicate keys
-// TODO: filter out the nouns into places and objects and animals
-// TODO: remove potential insensitive words
-func GetNatureThemeNounsData() []string {
-	data := map[string]bool{
-		"aurora":     true,
-		"bay":        true,
-		"beach":      true,
-		"bear":       true,
-		"bird":       true,
-		"boulder":    true,
-		"breeze":     true,
-		"brook":      true,
-		"bush":       true,
-		"butterfly":  true,
-		"cactus":     true,
-		"camel":      true,
-		"canopy":     true,
-		"canyon":     true,
-		"cave":       true,
-		"cliff":      true,
-		"cloud":      true,
-		"comet":      true,
-		"crater":     true,
-		"creek":      true,
-		"cygnet":     true,
-		"darkness":   true,
-		"dawn":       true,
-		"deer":       true,
-		"desert":     true,
-		"dew":        true,
-		"dream":      true,
-		"dune":       true,
-		"elegy":      true,
-		"feather":    true,
-		"fern":       true,
-		"field":      true,
-		"fire":       true,
-		"firefly":    true,
-		"flower":     true,
-		"fog":        true,
-		"forest":     true,
-		"fossil":     true,
-		"frog":       true,
-		"frost":      true,
-		"geyser":     true,
-		"glacier":    true,
-		"glade":      true,
-		"grass":      true,
-		"grotto":     true,
-		"grove":      true,
-		"haze":       true,
-		"hedge":      true,
-		"highland":   true,
-		"hill":       true,
-		"island":     true,
-		"ivy":        true,
-		"jungle":     true,
-		"knoll":      true,
-		"lagoon":     true,
-		"lake":       true,
-		"leaf":       true,
-		"lilypad":    true,
-		"marsh":      true,
-		"meadow":     true,
-		"moon":       true,
-		"morning":    true,
-		"moss":       true,
-		"mountain":   true,
-		"mushroom":   true,
-		"night":      true,
-		"noon":       true,
-		"oak":        true,
-		"ocean":      true,
-		"orchid":     true,
-		"pebble":     true,
-		"pine":       true,
-		"pond":       true,
-		"pumpkin":    true,
-		"rain":       true,
-		"ravine":     true,
-		"reef":       true,
-		"river":      true,
-		"root":       true,
-		"rose":       true,
-		"sand":       true,
-		"savanna":    true,
-		"sea":        true,
-		"seed":       true,
-		"shadow":     true,
-		"shark":      true,
-		"sky":        true,
-		"smoke":      true,
-		"snow":       true,
-		"snowflake":  true,
-		"soil":       true,
-		"squirrel":   true,
-		"star":       true,
-		"stone":      true,
-		"storm":      true,
-		"stream":     true,
-		"sun":        true,
-		"sunset":     true,
-		"surf":       true,
-		"swamp":      true,
-		"swan":       true,
-		"thunder":    true,
-		"tiger":      true,
-		"totem":      true,
-		"trail":      true,
-		"tree":       true,
-		"tundra":     true,
-		"turtle":     true,
-		"twilight":   true,
-		"valley":     true,
-		"violet":     true,
-		"water":      true,
-		"waterfall":  true,
-		"wave":       true,
-		"wilderness": true,
-		"willow":     true,
-		"wind":       true,
-		"wolf":       true,
-		"wood":       true,
-	}
-
-	return convertMapSetToStringArray(data)
-}
-
-// GetNatureThemeAdjectivesData
-// TODO: remove potential insensitive words
-func GetNatureThemeAdjectivesData() []string {
-	data := map[string]bool{
-		"aged":       true,
-		"alert":      true,
-		"ancient":    true,
-		"arctic":     true,
-		"arid":       true,
-		"autumn":     true,
-		"babbling":   true,
-		"barren":     true,
-		"billowing":  true,
-		"bitter":     true,
-		"blooming":   true,
-		"blue":       true,
-		"bold":       true,
-		"brave":      true,
-		"brisk":      true,
-		"broken":     true,
-		"callous":    true,
-		"celestial":  true,
-		"cold":       true,
-		"cool":       true,
-		"crimson":    true,
-		"crisp":      true,
-		"crumbling":  true,
-		"damp":       true,
-		"dark":       true,
-		"dazzling":   true,
-		"delicate":   true,
-		"dense":      true,
-		"diligent":   true,
-		"divine":     true,
-		"dry":        true,
-		"dull":       true,
-		"dusty":      true,
-		"earthy":     true,
-		"elegant":    true,
-		"empty":      true,
-		"evergreen":  true,
-		"fading":     true,
-		"falling":    true,
-		"floral":     true,
-		"forgotten":  true,
-		"fragrant":   true,
-		"full":       true,
-		"gloomy":     true,
-		"green":      true,
-		"hidden":     true,
-		"holy":       true,
-		"howling":    true,
-		"icy":        true,
-		"large":      true,
-		"lazy":       true,
-		"lingering":  true,
-		"little":     true,
-		"lively":     true,
-		"lonely":     true,
-		"long":       true,
-		"looming":    true,
-		"loud":       true,
-		"lush":       true,
-		"mild":       true,
-		"misty":      true,
-		"muddy":      true,
-		"nameless":   true,
-		"pallid":     true,
-		"patient":    true,
-		"polished":   true,
-		"pretty":     true,
-		"pristine":   true,
-		"proud":      true,
-		"purple":     true,
-		"quiet":      true,
-		"red":        true,
-		"restless":   true,
-		"rotting":    true,
-		"rustling":   true,
-		"rusty":      true,
-		"sacred":     true,
-		"shy":        true,
-		"silent":     true,
-		"silky":      true,
-		"sleeping":   true,
-		"slender":    true,
-		"small":      true,
-		"smooth":     true,
-		"solitary":   true,
-		"sparkling":  true,
-		"sparse":     true,
-		"spring":     true,
-		"still":      true,
-		"strong":     true,
-		"stunning":   true,
-		"summer":     true,
-		"timid":      true,
-		"tired":      true,
-		"towering":   true,
-		"tropical":   true,
-		"vast":       true,
-		"verdant":    true,
-		"vibrant":    true,
-		"wandering":  true,
-		"waning":     true,
-		"warm":       true,
-		"watchful":   true,
-		"weathered":  true,
-		"whispering": true,
-		"wild":       true,
-		"winter":     true,
-		"wispy":      true,
-		"yearning":   true,
-		"young":      true,
-		"zealous":    true,
-	}
-	return convertMapSetToStringArray(data)
-}
-
-func convertMapSetToStringArray(data map[string]bool) []string {
-	var arr []string
-	for k, _ := range data {
-		arr = append(arr, k)
-	}
-	return arr
-}
