+++ name_generator/adjectives.go
@@ -1,0 +1,124 @@
+package name_generator
+
+var adjectives []string
+var adjectivesSet = map[string]bool{
+	"aged":       true,
+	"alert":      true,
+	"ancient":    true,
+	"arctic":     true,
+	"arid":       true,
+	"autumn":     true,
+	"babbling":   true,
+	"barren":     true,
+	"billowing":  true,
+	"bitter":     true,
+	"blooming":   true,
+	"blue":       true,
+	"bold":       true,
+	"brave":      true,
+	"brisk":      true,
+	"broken":     true,
+	"callous":    true,
+	"celestial":  true,
+	"cold":       true,
+	"cool":       true,
+	"crimson":    true,
+	"crisp":      true,
+	"crumbling":  true,
+	"damp":       true,
+	"dark":       true,
+	"dazzling":   true,
+	"delicate":   true,
+	"dense":      true,
+	"diligent":   true,
+	"divine":     true,
+	"dry":        true,
+	"dull":       true,
+	"dusty":      true,
+	"earthy":     true,
+	"elegant":    true,
+	"empty":      true,
+	"evergreen":  true,
+	"fading":     true,
+	"falling":    true,
+	"floral":     true,
+	"forgotten":  true,
+	"fragrant":   true,
+	"full":       true,
+	"gloomy":     true,
+	"green":      true,
+	"hidden":     true,
+	"holy":       true,
+	"howling":    true,
+	"icy":        true,
+	"large":      true,
+	"lazy":       true,
+	"lingering":  true,
+	"little":     true,
+	"lively":     true,
+	"lonely":     true,
+	"long":       true,
+	"looming":    true,
+	"loud":       true,
+	"lush":       true,
+	"mild":       true,
+	"misty":      true,
+	"muddy":      true,
+	"nameless":   true,
+	"pallid":     true,
+	"patient":    true,
+	"polished":   true,
+	"pretty":     true,
+	"pristine":   true,
+	"proud":      true,
+	"purple":     true,
+	"quiet":      true,
+	"red":        true,
+	"restless":   true,
+	"rotting":    true,
+	"rustling":   true,
+	"rusty":      true,
+	"sacred":     true,
+	"shy":        true,
+	"silent":     true,
+	"silky":      true,
+	"sleeping":   true,
+	"slender":    true,
+	"small":      true,
+	"smooth":     true,
+	"solitary":   true,
+	"sparkling":  true,
+	"sparse":     true,
+	"spring":     true,
+	"still":      true,
+	"strong":     true,
+	"stunning":   true,
+	"summer":     true,
+	"timid":      true,
+	"tired":      true,
+	"towering":   true,
+	"tropical":   true,
+	"vast":       true,
+	"verdant":    true,
+	"vibrant":    true,
+	"wandering":  true,
+	"waning":     true,
+	"warm":       true,
+	"watchful":   true,
+	"weathered":  true,
+	"whispering": true,
+	"wild":       true,
+	"winter":     true,
+	"wispy":      true,
+	"yearning":   true,
+	"young":      true,
+	"zealous":    true,
+}
+
+// singleton pattern - where convertMapSetToStringArray will be called only once.
+func getAdjectives() []string {
+	once.Do(func() {
+		adjectives = convertMapSetToStringArray(adjectivesSet)
+	})
+	return adjectives
+}
