+++ name_generator/name_generator.go
@@ -7,33 +7,31 @@ import (
 	"time"
 )
 
-var randomNameGenerator *RandomNameGenerator
+var singletonRandomNameGenerator *randomNameGenerator
 var once sync.Once
 
-type RandomNameGenerator struct {
+type randomNameGenerator struct {
 	random *rand.Rand
 }
 
-type GeneratorArgs struct {
+type generatorArgs struct {
 	adjectives []string
 	nouns      []string
 }
 
-// GetNameGenerator we use once to seen random generator once
+// getNameGenerator we use once to seen random generator once
 // https://stackoverflow.com/questions/12321133/how-to-properly-seed-random-number-generator
 // has interesting read about maybe using crypto/rand in future if need be
-func GetNameGenerator() *RandomNameGenerator {
-	if randomNameGenerator == nil {
-		once.Do(func() {
-			seed := time.Now().UTC().UnixNano()
-			r := rand.New(rand.New(rand.NewSource(seed)))
-			randomNameGenerator = &RandomNameGenerator{random: r}
-		})
-	}
-	return randomNameGenerator
+func getNameGenerator() *randomNameGenerator {
+	once.Do(func() {
+		seed := time.Now().UTC().UnixNano()
+		r := rand.New(rand.New(rand.NewSource(seed)))
+		singletonRandomNameGenerator = &randomNameGenerator{random: r}
+	})
+	return singletonRandomNameGenerator
 }
 
-func (rng *RandomNameGenerator) GenerateName(args GeneratorArgs) string {
+func (rng *randomNameGenerator) generateName(args generatorArgs) string {
 	randomNoun := args.nouns[rng.random.Intn(len(args.nouns))]
 	randomAdjective := args.adjectives[rng.random.Intn(len(args.adjectives))]
