+++ name_generator/nature_theme_names_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func Test_generateNatureThemeNameForArtifactsInternal(t *testing.T) {
-	args := GeneratorArgs{
+	args := generatorArgs{
 		adjectives: []string{"test_adj", "test_adj_two"},
 		nouns:      []string{"noun"},
 	}
@@ -18,4 +18,18 @@ func Test_generateNatureThemeNameForArtifactsInternal(t *testing.T) {
 	}
 
 	require.Contains(t, potentialCandidates, actual)
+}
+
+func Test_convertMapSetToStringArray(t *testing.T) {
+	data := map[string]bool{
+		"test":             true,
+		"test_key_another": true,
+	}
+
+	actual := convertMapSetToStringArray(data)
+
+	require.Contains(t, actual, "test")
+	require.Contains(t, actual, "test_key_another")
+	require.NotContains(t, actual, "abc")
+	require.Len(t, actual, len(data))
 }
