+++ name_generator/nouns.go
@@ -1,0 +1,132 @@
+package name_generator
+
+var nouns []string
+var nounSet = map[string]bool{
+	"aurora":     true,
+	"bay":        true,
+	"beach":      true,
+	"bear":       true,
+	"bird":       true,
+	"boulder":    true,
+	"breeze":     true,
+	"brook":      true,
+	"bush":       true,
+	"butterfly":  true,
+	"cactus":     true,
+	"camel":      true,
+	"canopy":     true,
+	"canyon":     true,
+	"cave":       true,
+	"cliff":      true,
+	"cloud":      true,
+	"comet":      true,
+	"crater":     true,
+	"creek":      true,
+	"cygnet":     true,
+	"darkness":   true,
+	"dawn":       true,
+	"deer":       true,
+	"desert":     true,
+	"dew":        true,
+	"dream":      true,
+	"dune":       true,
+	"elegy":      true,
+	"feather":    true,
+	"fern":       true,
+	"field":      true,
+	"fire":       true,
+	"firefly":    true,
+	"flower":     true,
+	"fog":        true,
+	"forest":     true,
+	"fossil":     true,
+	"frog":       true,
+	"frost":      true,
+	"geyser":     true,
+	"glacier":    true,
+	"glade":      true,
+	"grass":      true,
+	"grotto":     true,
+	"grove":      true,
+	"haze":       true,
+	"hedge":      true,
+	"highland":   true,
+	"hill":       true,
+	"island":     true,
+	"ivy":        true,
+	"jungle":     true,
+	"knoll":      true,
+	"lagoon":     true,
+	"lake":       true,
+	"leaf":       true,
+	"lilypad":    true,
+	"marsh":      true,
+	"meadow":     true,
+	"moon":       true,
+	"morning":    true,
+	"moss":       true,
+	"mountain":   true,
+	"mushroom":   true,
+	"night":      true,
+	"noon":       true,
+	"oak":        true,
+	"ocean":      true,
+	"orchid":     true,
+	"pebble":     true,
+	"pine":       true,
+	"pond":       true,
+	"pumpkin":    true,
+	"rain":       true,
+	"ravine":     true,
+	"reef":       true,
+	"river":      true,
+	"root":       true,
+	"rose":       true,
+	"sand":       true,
+	"savanna":    true,
+	"sea":        true,
+	"seed":       true,
+	"shadow":     true,
+	"shark":      true,
+	"sky":        true,
+	"smoke":      true,
+	"snow":       true,
+	"snowflake":  true,
+	"soil":       true,
+	"squirrel":   true,
+	"star":       true,
+	"stone":      true,
+	"storm":      true,
+	"stream":     true,
+	"sun":        true,
+	"sunset":     true,
+	"surf":       true,
+	"swamp":      true,
+	"swan":       true,
+	"thunder":    true,
+	"tiger":      true,
+	"totem":      true,
+	"trail":      true,
+	"tree":       true,
+	"tundra":     true,
+	"turtle":     true,
+	"twilight":   true,
+	"valley":     true,
+	"violet":     true,
+	"water":      true,
+	"waterfall":  true,
+	"wave":       true,
+	"wilderness": true,
+	"willow":     true,
+	"wind":       true,
+	"wolf":       true,
+	"wood":       true,
+}
+
+// singleton pattern - where convertMapSetToStringArray will be called only once.
+func getNouns() []string {
+	once.Do(func() {
+		nouns = convertMapSetToStringArray(nounSet)
+	})
+	return nouns
+}
